% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dicionários}
\alias{Dicionários}
\alias{vinculo}
\alias{cor}
\alias{instrucao}
\alias{uf}
\title{Dicionário de variáveis categóricas}
\usage{
vinculo

cor

instrucao

uf
}
\description{
Dicionários com par código/descrição.
}
\keyword{datasets}
