% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dados}
\alias{Dados}
\alias{moto1}
\alias{moto2}
\title{Motoboys na PNAD contítnua}
\format{
Uma tibble com 11.044 linhas e 8 variáveis (moto1)
ou uma tibble com 686 linhas e 8 variáveis (moto2):
\describe{
\item{ano}{Ano da PNAD. Todos os dados são da PNAD do primeiro trimestre.}
\item{uf}{Unidade da Federação. Para códigos ver \code{\link{uf}}.}
\item{sexo}{Sexo. 1 = homem, 2 = mulher.}
\item{idade}{Idade, em anos.}
\item{cor}{Cor. Para códigos ver \code{\link{cor}}.}
\item{vinculo}{Vínculo no mercado de trbaalho. Para códigos ver \code{\link{vinculo}}.}
\item{renda}{Renda, em reais correntes. Ver  \code{\link{ipca}} para deflacionar}
\item{jornada}{Jornada de trabalho semanal, em horas.}
\item{instrucao}{Instrução. Para códigos ver \code{\link{instrucao}}.}
}
}
\usage{
moto1

moto2
}
\description{
Esses dados perderam o vinculo com o desenho da amostra e
tem propósito puramente educacional.

Os dados de \code{moto1} se referem as estatísticas sobre vinculo principal
no mercado de trabalho enquanto \code{moto2} aos casos secundários.
}
\examples{
head(moto1)
head(moto2)
}
\keyword{datasets}
