% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dicionários}
\alias{Dicionários}
\alias{vinculo}
\alias{cor}
\alias{instrucao}
\alias{uf}
\alias{ipca}
\title{Dicionário de variáveis categóricas}
\format{
NULL
\describe{
\item{ano}{Ano.}
\item{ipca}{Variação anual do IPCA, em percentual.}
\item{deflator}{Deflator usado para reais correntes em reais de 2020.}
}
}
\usage{
vinculo

cor

instrucao

uf

ipca
}
\description{
Dicionários com par código/descrição.

Variação anual do IPCA e deflator para valores 2020
}
\keyword{datasets}
